Move content of test_identifiers inside test_article
